Some clean-up and warning fixes

+++ src/interp/lex.c
@@ -210,7 +210,7 @@ lextok (PACTX pactx, YYSTYPE * lvalp, INT c, INT t)
 		p = tokbuf;
 		while (TRUE) {
 			while ((c = inchar(pactx)) != EOF && c != '"' && c != '\\') {
-				if (p-tokbuf > sizeof(tokbuf)/sizeof(tokbuf[0]) - 3) {
+				if (p-tokbuf > (int)(sizeof(tokbuf)/sizeof(tokbuf[0]) - 3)) {
 					/* Overflowing tokbuf buffer */
 					/* TODO: (Perry, 2006-06-30) I don't know how to fail gracefully from here inside parser */
 					char msg[512];
